+++ apps/api/src/app/user/interfaces/user-settings-params.interface.ts
@@ -1,7 +1,0 @@
-import { ViewMode } from '@prisma/client';
-
-export interface UserSettingsParams {
-  currency?: string;
-  userId: string;
-  viewMode?: ViewMode;
-}