+++ apps/api/src/app/user/interfaces/user-settings.interface.ts
@@ -1,5 +1,0 @@
-export interface UserSettings {
-  emergencyFund?: number;
-  locale?: string;
-  isRestrictedView?: boolean;
-}