+++ apps/api/src/app/user/update-user-setting.dto.ts
@@ -1,6 +1,22 @@
-import { IsBoolean, IsNumber, IsOptional, IsString } from 'class-validator';
+import type { DateRange } from '@ghostfolio/common/types';
+import { ViewMode } from '@prisma/client';
+import {
+  IsBoolean,
+  IsIn,
+  IsNumber,
+  IsOptional,
+  IsString
+} from 'class-validator';
 
 export class UpdateUserSettingDto {
+  @IsOptional()
+  @IsString()
+  baseCurrency?: string;
+
+  @IsIn(<DateRange[]>['1d', '1y', '5y', 'max', 'ytd'])
+  @IsOptional()
+  dateRange?: DateRange;
+
   @IsNumber()
   @IsOptional()
   emergencyFund?: number;
@@ -24,4 +40,8 @@ export class UpdateUserSettingDto {
   @IsNumber()
   @IsOptional()
   savingsRate?: number;
+
+  @IsIn(<ViewMode[]>['DEFAULT', 'ZEN'])
+  @IsOptional()
+  viewMode?: ViewMode;
 }