+++ apps/api/src/models/interfaces/user-settings.interface.ts
@@ -1,3 +1,0 @@
-export interface UserSettings {
-  baseCurrency: string;
-}